* commit 'a79399506029f0e642dba96f7b63163655b72be6':
  remove debug lines
  Parity with metadataAvailabilityDate

+++ transformer/src/main/java/com/netflix/vms/transformer/hollowinput/VideoGeneralDelegate.java
@@ -32,9 +32,9 @@ public interface VideoGeneralDelegate extends HollowObjectDelegate {
 
     public int getOriginalLanguageBcpCodeOrdinal(int ordinal);
 
-    public long getMetadataReleaseDays(int ordinal);
+    public int getMetadataReleaseDays(int ordinal);
 
-    public Long getMetadataReleaseDaysBoxed(int ordinal);
+    public Integer getMetadataReleaseDaysBoxed(int ordinal);
 
     public int getOriginCountryCodeOrdinal(int ordinal);
